+++ src/coord/src/catalog/config.rs
@@ -28,6 +28,10 @@ pub struct Config<'a> {
     pub enable_logging: bool,
     /// Information about this build of Materialize.
     pub build_info: &'static BuildInfo,
+    /// An [External ID][] to use for all AWS AssumeRole operations.
+    ///
+    /// [External ID]: https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_create_for-user_externalid.html
+    pub aws_external_id: Option<String>,
     /// Timestamp frequency to use for CREATE SOURCE
     pub timestamp_frequency: Duration,
     /// Function to generate wall clock now; can be mocked.