+++ src/coord/src/coord.rs
@@ -108,7 +108,7 @@ use tokio::select;
 use tokio::sync::{mpsc, oneshot, watch};
 
 use build_info::BuildInfo;
-use dataflow_types::client::{ComputeClient, StorageClient};
+use dataflow_types::client::{ComputeClient, StorageClient, DEFAULT_COMPUTE_INSTANCE_ID};
 use dataflow_types::client::{ComputeResponse, TimestampBindingFeedback};
 use dataflow_types::client::{Response as DataflowResponse, StorageResponse};
 use dataflow_types::logging::LoggingConfig as DataflowLoggingConfig;
@@ -256,6 +256,7 @@ pub struct Config {
     pub disable_user_indexes: bool,
     pub safe_mode: bool,
     pub build_info: &'static BuildInfo,
+    pub aws_external_id: Option<String>,
     pub metrics_registry: MetricsRegistry,
     pub persister: PersisterWithConfig,
     pub now: NowFn,
@@ -732,6 +733,11 @@ impl Coordinator {
                 }
             });
         }
+
+        // Create the default cluster so that subsequent default cluster commands succeed.
+        self.dataflow_client
+            .create_instance(DEFAULT_COMPUTE_INSTANCE_ID)
+            .await;
 
         let mut metric_scraper_stream = self.metric_scraper.tick_stream();
 
@@ -1480,7 +1486,10 @@ impl Coordinator {
             // well.
             let item = self.catalog.try_get_by_id(*id).map(|e| e.item());
             if let Some(CatalogItem::Index(catalog::Index { on, .. })) = item {
-                if let Some(persist) = self.persister.table_details.get(&id) {
+                // We only want to forward since updates to persist if they:
+                //  - are from the primary index on a table
+                //  - and that table itself is currently persisted
+                if let Some(persist) = self.persister.table_details.get(&on) {
                     if self.catalog.default_index_for(*on) == Some(*id) {
                         table_since_updates.push((persist.stream_id, frontier.clone()));
                         table_source_since_updates.push((*on, frontier.clone()));
@@ -1555,7 +1564,7 @@ impl Coordinator {
             // See #10300 for context.
             self.persisted_table_allow_compaction(&index_since_updates);
             self.dataflow_client
-                .allow_index_compaction(index_since_updates)
+                .allow_index_compaction(DEFAULT_COMPUTE_INSTANCE_ID, index_since_updates)
                 .await;
         }
 
@@ -1688,7 +1697,9 @@ impl Coordinator {
             let _ = conn_meta.cancel_tx.send(Canceled::Canceled);
 
             // Allow dataflow to cancel any pending peeks.
-            self.dataflow_client.cancel_peek(conn_id).await;
+            self.dataflow_client
+                .cancel_peek(DEFAULT_COMPUTE_INSTANCE_ID, conn_id)
+                .await;
         }
     }
 
@@ -4113,7 +4124,9 @@ impl Coordinator {
                 for id in sinks_to_drop.iter() {
                     self.sink_writes.remove(id);
                 }
-                self.dataflow_client.drop_sinks(sinks_to_drop).await;
+                self.dataflow_client
+                    .drop_sinks(DEFAULT_COMPUTE_INSTANCE_ID, sinks_to_drop)
+                    .await;
             }
             if !indexes_to_drop.is_empty() {
                 self.drop_indexes(indexes_to_drop).await;
@@ -4198,7 +4211,9 @@ impl Coordinator {
 
     async fn drop_sinks(&mut self, dataflow_names: Vec<GlobalId>) {
         if !dataflow_names.is_empty() {
-            self.dataflow_client.drop_sinks(dataflow_names).await;
+            self.dataflow_client
+                .drop_sinks(DEFAULT_COMPUTE_INSTANCE_ID, dataflow_names)
+                .await;
         }
     }
 
@@ -4210,7 +4225,9 @@ impl Coordinator {
             }
         }
         if !trace_keys.is_empty() {
-            self.dataflow_client.drop_indexes(trace_keys).await
+            self.dataflow_client
+                .drop_indexes(DEFAULT_COMPUTE_INSTANCE_ID, trace_keys)
+                .await
         }
     }
 
@@ -4324,7 +4341,9 @@ impl Coordinator {
         for dataflow in dataflows.into_iter() {
             dataflow_plans.push(self.finalize_dataflow(dataflow));
         }
-        self.dataflow_client.create_dataflows(dataflow_plans).await;
+        self.dataflow_client
+            .create_dataflows(DEFAULT_COMPUTE_INSTANCE_ID, dataflow_plans)
+            .await;
     }
 
     /// Finalizes a dataflow.
@@ -4557,6 +4576,7 @@ pub async fn serve(
         disable_user_indexes,
         safe_mode,
         build_info,
+        aws_external_id,
         metrics_registry,
         persister,
         now,
@@ -4571,6 +4591,7 @@ pub async fn serve(
         safe_mode,
         enable_logging: logging.is_some(),
         build_info,
+        aws_external_id,
         timestamp_frequency,
         now: now.clone(),
         skip_migrations: false,
@@ -4622,14 +4643,17 @@ pub async fn serve(
             };
             if let Some(config) = &logging {
                 handle.block_on(
-                    coord.dataflow_client.enable_logging(DataflowLoggingConfig {
-                        granularity_ns: config.granularity.as_nanos(),
-                        active_logs: BUILTINS
-                            .logs()
-                            .map(|src| (src.variant.clone(), src.index_id))
-                            .collect(),
-                        log_logging: config.log_logging,
-                    }),
+                    coord.dataflow_client.enable_logging(
+                        DEFAULT_COMPUTE_INSTANCE_ID,
+                        DataflowLoggingConfig {
+                            granularity_ns: config.granularity.as_nanos(),
+                            active_logs: BUILTINS
+                                .logs()
+                                .map(|src| (src.variant.clone(), src.index_id))
+                                .collect(),
+                            log_logging: config.log_logging,
+                        },
+                    ),
                 );
             }
             let bootstrap = handle.block_on(coord.bootstrap(builtin_table_updates));
@@ -4842,7 +4866,7 @@ fn check_statement_safety(stmt: &Statement<Raw>) -> Result<(), CoordError> {
 /// or by reading out of existing arrangements, and implements the appropriate plan.
 pub mod fast_path_peek {
 
-    use dataflow_types::client::ComputeClient;
+    use dataflow_types::client::{ComputeClient, DEFAULT_COMPUTE_INSTANCE_ID};
     use std::collections::HashMap;
 
     use crate::CoordError;
@@ -5018,7 +5042,9 @@ pub mod fast_path_peek {
                     thinned_arity: index_thinned_arity,
                 }) => {
                     // Very important: actually create the dataflow (here, so we can destructure).
-                    self.dataflow_client.create_dataflows(vec![dataflow]).await;
+                    self.dataflow_client
+                        .create_dataflows(DEFAULT_COMPUTE_INSTANCE_ID, vec![dataflow])
+                        .await;
                     // Create an identity MFP operator.
                     let mut map_filter_project = expr::MapFilterProject::new(source_arity);
                     map_filter_project
@@ -5058,6 +5084,7 @@ pub mod fast_path_peek {
             let (id, key, conn_id, timestamp, _finishing, map_filter_project) = peek_command;
             self.dataflow_client
                 .peek(
+                    DEFAULT_COMPUTE_INSTANCE_ID,
                     id,
                     key,
                     conn_id,