+++ src/dataflow/src/sink/kafka.rs
@@ -87,7 +87,7 @@ where
         sink_id: GlobalId,
         sinked_collection: Collection<G, (Option<Row>, Option<Row>), Diff>,
         metrics: &SinkBaseMetrics,
-    ) -> Option<Box<dyn Any>>
+    ) -> Option<Rc<dyn Any>>
     where
         G: Scope<Timestamp = Timestamp>,
     {
@@ -950,7 +950,7 @@ fn kafka<G>(
     as_of: SinkAsOf,
     write_frontier: Rc<RefCell<Antichain<Timestamp>>>,
     metrics: &KafkaBaseMetrics,
-) -> Box<dyn Any>
+) -> Rc<dyn Any>
 where
     G: Scope<Timestamp = Timestamp>,
 {
@@ -1027,7 +1027,7 @@ pub fn produce_to_kafka<G>(
     shared_gate_ts: Rc<Cell<Option<Timestamp>>>,
     write_frontier: Rc<RefCell<Antichain<Timestamp>>>,
     metrics: &KafkaBaseMetrics,
-) -> Box<dyn Any>
+) -> Rc<dyn Any>
 where
     G: Scope<Timestamp = Timestamp>,
 {
@@ -1287,7 +1287,7 @@ where
         }),
     );
 
-    Box::new(KafkaSinkToken { shutdown_flag })
+    Rc::new(KafkaSinkToken { shutdown_flag })
 }
 
 /// Encodes a stream of `(Option<Row>, Option<Row>)` updates using the specified encoder.